+++ src/utils/number/index.js
@@ -1,3 +1,3 @@
 export * from './currencyFormat';
 export * from './numberFormat';
-export * from './nHashtoHash';
+export * from './formatDenom';